Move IDs are now a byte instead of an int. Should reduce the memory footprint of the system slightly.

+++ Libraries/Foat/Puzzles/RubiksCube/RubiksCube.Moves.cs
@@ -50,7 +50,7 @@
             BackHalf.OppositeMove = BackHalf;
         }
 
-        public static Move<RubiksCube> GetMove(int moveId)
+        public static Move<RubiksCube> GetMove(byte moveId)
         {
             if (moveId < 0 || moveId > AllMoves.Length)
                 throw new ArgumentException("moveId out of range.");
