+++ Libraries/Foat/Puzzles/Solutions/PuzzleFactory.cs
@@ -6,6 +6,6 @@ using System;
     internal static class PuzzleFactory<TPuzzle> where TPuzzle : IPuzzle<TPuzzle>, IEquatable<TPuzzle>
     {
         internal static Func<byte[], TPuzzle> PuzzleFromBytes { get; set; }
-        internal static Func<int, Move<TPuzzle>> FindMove { get; set; }
+        internal static Func<byte, Move<TPuzzle>> FindMove { get; set; }
     }
 }
